+++ src/components/shared-spatial-scatterplot/AbstractSpatialOrScatterplot.js
@@ -128,6 +128,10 @@ export default class AbstractSpatialOrScatterplot extends PureComponent {
 
     const showCellSelectionTools = this.cellsLayer !== null;
     const showPanTool = this.cellsLayer !== null;
+    // For large datasets, the visual quality takes only a small
+    // hit in exchange for much better performance by setting this to false:
+    // https://deck.gl/docs/api-reference/core/deck#usedevicepixels
+    const useDevicePixels = this.cellsEntries.length < 100000;
 
     return (
       <>
@@ -149,6 +153,7 @@ export default class AbstractSpatialOrScatterplot extends PureComponent {
           onWebGLInitialized={this.onWebGLInitialized}
           onViewStateChange={this.onViewStateChange}
           viewState={viewState}
+          useDevicePixels={useDevicePixels}
           controller={tool ? ({ dragPan: false }) : true}
           getCursor={tool ? getCursorWithTool : getCursor}
         >